fix(customer): relate customers directly to shipper instead of preferred shipping
- Replace preferred_shipping_id with shipper_id across Customer model, UpdateCustomer, EditCustomer, and Retina customer actions
- Add Shipper::preferredShippings relation and validate shipper_id exists via preferred_shippings scoped to shop

+++ database/migrations/2026_07_21_071428_add_shipper_id_to_customers_table.php
@@ -14,7 +14,7 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->unsignedInteger('shipper_id')->nullable()->index();
+            $table->unsignedSmallInteger('shipper_id')->nullable()->index();
             $table->foreign('shipper_id')->references('id')->on('shippers')->nullOnDelete();
         });
     }
